extract common code of blog, calendar generation
