+++ lib/generate-blog-components/index.js
@@ -5,14 +5,14 @@ const path = require('path');
 
 const writeFile = require('broccoli-file-creator');
 const mergeTrees = require('broccoli-merge-trees');
-const marked = require('marked');
 const highlightjs = require('highlight.js');
 const dateformat = require('dateformat');
-const jsdom = require('jsdom');
+const marked = require('marked');
 const handlebars = require('handlebars');
 const fs = require('fs-extra');
 
 const collectPosts = require('./lib/collect-posts');
+const htmlizeMarkdown = require('../markdown-content/htmlize-markdown');
 
 const Renderer = new marked.Renderer();
 
@@ -126,12 +126,7 @@ function encodeCurlies(content) {
 }
 
 function htmlizeBody(body) {
-  let html = marked(body, { renderer: Renderer });
-  html = manipulateDom(html, dom => {
-    dom.querySelectorAll('h2').forEach(h2 => h2.classList.add('typography.headline-2'));
-    dom.querySelectorAll('h3').forEach(h3 => h3.classList.add('typography.headline-3'));
-    dom.querySelectorAll('h4').forEach(h4 => h4.classList.add('typography.headline-4'));
-    dom.querySelectorAll('p').forEach(p => p.classList.add('typography.body-text'));
+  let html = htmlizeMarkdown(body, { renderer: Renderer }, dom => {
     dom.querySelectorAll('p > img').forEach(img => {
       let paragraph = img.parentElement;
       img.classList.add('main-block.image-centered');
@@ -141,35 +136,18 @@ function htmlizeBody(body) {
       paragraph.parentElement.insertBefore(div, paragraph);
       paragraph.parentElement.removeChild(paragraph);
     });
-    dom.querySelectorAll('ul, ol').forEach(list => list.classList.add('list'));
-    dom.querySelectorAll('ul > li').forEach(li => li.classList.add('list.item'));
-    dom.querySelectorAll('ol > li').forEach(li => li.classList.add('list.item-ordered'));
-    dom.querySelectorAll('table').forEach(table => table.classList.add('table'));
-    dom.querySelectorAll('th').forEach(th => th.classList.add('table.header'));
-    dom.querySelectorAll('td').forEach(td => td.classList.add('table.cell'));
-    dom.querySelectorAll('a').forEach(a => {
-      if (!a.href.startsWith('http')) {
-        a.dataset.internal = true;
-      }
-      a.classList.add('typography.link');
-    });
   });
   return encodeCurlies(html);
 }
 
 function htmlizeExcerpt(excerpt) {
-  let html = marked(excerpt, { renderer: Renderer });
-  html = manipulateDom(html, dom => {
-    dom.querySelectorAll('p').forEach(p => p.classList.add('typography.lead'));
-    dom.querySelectorAll('a').forEach(a => a.classList.add('typography.link'));
+  let html = htmlizeMarkdown(excerpt, { renderer: Renderer }, dom => {
+    dom.querySelectorAll('p').forEach(p => {
+      p.classList.remove('typography.body-text');
+      p.classList.add('typography.lead');
+    });
   });
   return encodeCurlies(html);
-}
-
-function manipulateDom(html, callback) {
-  let dom = new jsdom.JSDOM(html);
-  callback(dom.window.document);
-  return dom.window.document.body.innerHTML;
 }
 
 function prepareTemplates() {
