+++ lib/generate-calendar/index.js
@@ -5,12 +5,11 @@ const path = require('path');
 const fs = require('fs-extra');
 const writeFile = require('broccoli-file-creator');
 const mergeTrees = require('broccoli-merge-trees');
-const marked = require('marked');
 const handlebars = require('handlebars');
-const jsdom = require('jsdom');
 const dateformat = require('dateformat');
 
 const collectEvents = require('./lib/collect-events');
+const htmlizeMarkdown = require('../markdown-content/htmlize-markdown');
 
 module.exports = {
   name: require('./package').name,
@@ -34,13 +33,13 @@ module.exports = {
   _writePageComponentTree(events) {
     let trees = [];
 
-    let futureEvents = sortClosestFirst(filterFutureEventsOnly(events)).map((event) => {
+    let futureEvents = sortClosestFirst(filterFutureEventsOnly(events)).map(event => {
       return {
         meta: {
           ...event.meta,
           date: dateformat(event.meta.date, 'mmmm d, yyyy'),
         },
-        content: htmlizeBody(event.content),
+        content: htmlizeMarkdown(event.content),
       };
     });
     let [futureConferences, futureMeetups] = splitConferencesMeetups(futureEvents);
@@ -59,7 +58,7 @@ module.exports = {
 };
 
 function filterFutureEventsOnly(events) {
-  return events.filter((event) => event.meta.date > new Date());
+  return events.filter(event => event.meta.date > new Date());
 }
 
 function sortClosestFirst(events) {
@@ -67,35 +66,15 @@ function sortClosestFirst(events) {
 }
 
 function splitConferencesMeetups(futureEvents) {
-  let conferences = futureEvents.filter((event) => event.meta.kind === 'conference');
-  let meetups = futureEvents.filter((event) => event.meta.kind === 'meetup');
+  let conferences = futureEvents.filter(event => event.meta.kind === 'conference');
+  let meetups = futureEvents.filter(event => event.meta.kind === 'meetup');
 
   return [conferences, meetups];
 }
 
-function htmlizeBody(body) {
-  let html = marked(body);
-  html = manipulateDom(html, dom => {
-    dom.querySelectorAll('p').forEach(p => p.classList.add('typography.body-text'));
-    dom.querySelectorAll('a').forEach(a => {
-      if (!a.href.startsWith('http')) {
-        a.dataset.internal = true;
-      }
-      a.classList.add('typography.link');
-    });
-  });
-  return html;
-}
-
-function manipulateDom(html, callback) {
-  let dom = new jsdom.JSDOM(html);
-  callback(dom.window.document);
-  return dom.window.document.body.innerHTML;
-}
-
 function prepareTemplates() {
   let templates = {
-    'template': ['component-template.hbs'],
+    template: ['component-template.hbs'],
     'css-block': ['component-css-block.hbs'],
   };
   return Object.keys(templates).reduce((acc, key) => {
