+++ lib/generate-calendar/lib/collect-events.js
@@ -1,8 +1,6 @@
 /* eslint-env node */
 'use strict';
 
-const path = require('path');
-
 const glob = require('glob');
 
 const separateMeta = require('../../markdown-content/separate-meta');
