+++ lib/generate-calendar/lib/files/page/component-css-block.hbs
