+++ lib/generate-calendar/lib/files/page/component-template.hbs
@@ -34,7 +34,9 @@
                     {{meta.title}}
                   </h3>
                   <p class="typography.body-text">
-                    {{meta.date}}, {{meta.location}}
+                    {{meta.date}}
+                    ,
+                    {{meta.location}}
                   </p>
                   {{#if content}}
                     {{{content}}}
@@ -64,7 +66,9 @@
                     {{meta.title}}
                   </h3>
                   <p class="typography.body-text">
-                    {{meta.date}}, {{meta.location}}
+                    {{meta.date}}
+                    ,
+                    {{meta.location}}
                   </p>
                   {{#if content}}
                     {{{content}}}
